ws-server/requirements.txt
python-dotenv==1.0.1
asyncio==3.4.3
websockets==14.1
